Move pytest option "--pipe-run-mode" to shared_pytest_plugins.py
